Add draw logo by Python and Matlab
The enology server we used for logo drawing is done for a while. We add
the logo drawing by Python and Matlab instead.
However, the python script on our server still has issues due to aging
hardware. The server is running but without logo drawing. We will solve
that issue asap. Stand-alone app is recommend for now.